Added basic bot workflow
The basic functionality for our event handler is working such that we
can listen for events that we're subscribed to in Slack. Currently, we
have the paths in place to ensure that we respond correctly to the
"challenge" query from Slack to verify our endpoint URL, ensure that
Slack events that we do get are from Slack itself, as well as processing
message events.

These message events are passed into our event handler method which
decides which events to pass to our bot. The bot is pretty simple right
now, only echoing the messages that it sees in the channels that it's
in.