Add karma functionality
We added an Item class to the project which stores the number of pluses
and minuses that an Item has received and a computed property displaying
the total score of an Item. We also added regular expressions to the
EventManager class to determine if a message that the bot sees is
something it needs to act upon. Added increment and decrement methods to
the bot class.