Fixed a few bugs
I forgot to import the Item class into the bot class as well as
prepending the class methods with self. There was also an issue where
the was no guarantee that the event detail would have a text property,
so I started using get() to get either the text property or an empty
string if nothing is there. Also added a host parameter to make sure
that the service is bound to the IP address of the actual host.