Added some framework for importing/exporting karma
Added some code to allow for the importing and exporting of karma
statuses via JSON. On initial start of the app, it will look for a JSON
file that contains the karma data (which is passed in as an environment
variable on the command line) and will populate it's internal dictionary
with that data.

Also fixed some PEP8 issues that came up.