Added ability to save and load karma to JSON file
One of the things that was critically missing from the minimum viable
product with KarmaChameleon was the ability for the karma information to
be saved across restarts of the bot. This commit adds the ability to
load from an existing JSON file which contains the karma data.

Also added was the functionality to save the karma data every time that
it's changed. Now the JSON file is updated every time an increment or
decrement operation occurs. Currently it will only work for a single
Slack workspace as it does not persist the name of the team in the JSON.
This could be done as a JSON object which holds the name (or ID) of the
Slack team and then the list of Karma items associated with that team.

The saving of events on each event also might not scale if this project
is ever put on the Slack Marketplace. It will likely work fine for a
small team but multiple teams might cause trouble. Some possible
solutions might be to have a thread spawned that checks for changes in
the karma dictionary every so often (say 15 minutes) and will save out
the in-memory representation of the karma info to the JSON file if
needed. Another possible solution might be to save everything into a
small database which would store all of the data for multiple teams and
their individual karma scores.