Add support for whitespace in increment and decrement messages
Before, the regexes were all constructed such that the operators (the ++
and --) had to be placed adjacent to the thing that was getting operated
on. This patch updates the regexes to be tolerant of a single whitespace
character so that messages with a space between the item and the
operator are counted as valid operations. This is handy because username
completion in Slack will, by default, leave a space after completing the
username.