Update to Python 3.11, small refactorings, remove Pandas
This patch moves the ceiling of which Python version KarmaChameleon
supports up to 3.11 (though testing was done locally with 3.10.7). This
also allowed us to update the library dependencies to the latest
available and fix any issues that came up with the updates to the
linting tools.

I also took the liberty of taking out the Pandas dependency from the
leaderboard functionality. Yes, this breaks the leaderboard, but there's
probably a lighter-weight way of doing what Pandas was doing without the
overhead.