Update to Python 3.11, small refactorings, remove Pandas (#54)
* Update to Python 3.11, small refactorings, remove Pandas

This patch moves the ceiling of which Python version KarmaChameleon
supports up to 3.11 (though testing was done locally with 3.10.7). This
also allowed us to update the library dependencies to the latest
available and fix any issues that came up with the updates to the
linting tools.

* Rebase on top of main, fix the fallout of that.

Remember to `git pull` BEFORE you start working, folks!