Use Pip Compile for Dependency Management (#58)
Previously, we were specifying minimum versions for our dependencies
and would just allow the latest versions of those packages to be
installed. However, this can be dangerous as we might install something
too new that might have a breaking change. This change moves to using
`pip-compile` for dependencies to make the builds and releases more
repeatable. Heck, it might even give dependabot something to do!